% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{cuda_home}
\alias{cuda_home}
\title{Installation Directory of the CUDA Components}
\usage{
cuda_home()
}
\value{
(\code{character(1)})\cr
Path to the installation directory (which may not exist yet).
}
\description{
Returns the directory into which the CUDA components are downloaded.
This is \code{tools::R_user_dir("pjrt.cuda", "cache")} unless the environment
variable \code{PJRT_CUDA_HOME} is set.
}
