% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{cuda_installed}
\alias{cuda_installed}
\title{Are the CUDA Components Installed?}
\usage{
cuda_installed()
}
\value{
(\code{logical(1)})
}
\description{
Checks whether the CUDA components selected by \code{PJRT_CUDA_COMPONENTS}
(default: all) are installed in \code{\link[=cuda_home]{cuda_home()}} with the versions pinned by
this version of the package.
}
