% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{lib_path}
\alias{lib_path}
\title{Path to the Shared Library Directory}
\usage{
lib_path()
}
\value{
(\code{character(1)})\cr
Path to the lib directory.
}
\description{
All CUDA component shared libraries are installed into a single directory,
which this function returns.
}
